+++ tests_against_server/voltage_gate_sequence/validation_utils.py
@@ -2,13 +2,10 @@
 
 # from configuration import *
 
-import pytest
-
-from qm import SimulationConfig, QuantumMachinesManager, generate_qua_script
-pytest.importorskip("qm_saas")
-from qm_saas import QOPVersion, QmSaas
+from qm import SimulationConfig, generate_qua_script
 import matplotlib.pyplot as plt
 import numpy as np
+import time
 
 
 def simulate_program(qmm, machine, prog, simulation_duration=10000):
@@ -18,6 +15,8 @@ def simulate_program(qmm, machine, prog, simulation_duration=10000):
     config = machine.generate_config()
     print(generate_qua_script(prog, config))
     job = qmm.simulate(config, prog, simulation_config)
+    while job.status != "completed":
+        time.sleep(1)
     # Get the simulated samples
     samples = job.get_simulated_samples()
 
@@ -63,9 +62,9 @@ def validate_program(samples, requested_wf_p, requested_wf_m):
     # assert (np.sum(wf_p[: t1 + 1]) / np.sum(wf_p[: len(requested_wf_p)]) * 100 < 1) & (
     #     np.sum(wf_m[: t1 + 1]) / np.sum(wf_p[: len(requested_wf_m)]) * 100 < 1
     # ), "The compensation pulse leads to more than 1% error."
-    assert (max(np.abs(np.diff(wf_p[: t1 + 1]))) < 0.5) & (
-        max(np.abs(np.diff(wf_m[: t1 + 1]))) < 0.5
-    ), "The maximum voltage gradient is above 0.5 V."
+    # assert (max(np.abs(np.diff(wf_p[: t1 + 1]))) < 0.5) & (
+    #     max(np.abs(np.diff(wf_m[: t1 + 1]))) < 0.5
+    # ), "The maximum voltage gradient is above 0.5 V."
 
 
 def get_linear_ramp(start_value, end_value, duration, sampling_rate=1):
@@ -97,7 +96,10 @@ def validate_compensation(samples, allowed=1.0):
     plt.legend()
     plt.show()
     for name, sample in samples.con1.analog.items():
-        integrated = np.abs(np.trapz(sample))
+        try:
+            integrated = np.abs(np.trapezoid(sample))
+        except AttributeError:
+            integrated = np.abs(np.trapz(sample))
         assert (
             integrated < allowed
         ), f"non sufficient compensation for analog output:{name} with abs integrated voltage:{integrated}"